requirements.txt
pypdf
transformers
torch
accelerate
pyyaml
pytest
pandas
